requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
inflect==2.1.0
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==4.4.2
Jinja2==2.10.1
lazy-object-proxy==1.4.2
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
pylint==2.3.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.6
wikipedia==1.4.0
wolframalpha==3.0.1
wrapt==1.11.2
xmltodict==0.12.0
